requirements_airflow.txt
pydantic
SQLAlchemy
typing_extensions
wheel
psycopg2-binary
pymongo
apache-airflow
apache-airflow-providers-postgres